split new datasets code and separate UI from business logic
