+++ renku/core/commands/remove.py
@@ -25,11 +25,11 @@ from renku.core import errors
 from renku.core.management.command_builder import inject
 from renku.core.management.command_builder.command import Command
 from renku.core.management.dataset.datasets_provenance import DatasetsProvenance
-from renku.core.management.dataset.usecase import remove_file
 from renku.core.management.interface.client_dispatcher import IClientDispatcher
 from renku.core.management.interface.dataset_gateway import IDatasetGateway
 from renku.core.utils import communication
 from renku.core.utils.git import get_git_user
+from renku.core.utils.os import delete_file
 
 
 @inject.autoparams()
@@ -70,7 +70,7 @@ def _remove(sources, edit_command, client_dispatcher: IClientDispatcher, dataset
                 dataset = dataset.copy()
                 for key in remove:
                     dataset.unlink_file(key)
-                    remove_file(client.path / key)
+                    delete_file(client.path / key, follow_symlinks=True)
 
                 datasets_provenance = DatasetsProvenance()
                 datasets_provenance.add_or_update(dataset, creator=get_git_user(client.repository))
