--- renku/core/management/dataset/usecase.py
@@ -1,934 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Copyright 2017-2021 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
-# Eidgenössische Technische Hochschule Zürich (ETHZ).
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-"""Dataset business logic."""
-
-
-import concurrent.futures
-import glob
-import os
-import shutil
-import time
-import urllib
-import uuid
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Set
-
-from renku.core import errors
-from renku.core.management.command_builder.command import inject
-from renku.core.management.dataset.constant import (
-    CACHE,
-    SUPPORTED_SCHEMES,
-    renku_dataset_images_path,
-    renku_pointers_path,
-)
-from renku.core.management.dataset.datasets_provenance import DatasetsProvenance
-from renku.core.management.dataset.request_model import ImageRequestModel
-from renku.core.management.interface.client_dispatcher import IClientDispatcher
-from renku.core.management.interface.dataset_gateway import IDatasetGateway
-from renku.core.metadata.immutable import DynamicProxy
-from renku.core.metadata.repository import Repository
-from renku.core.models.dataset import Dataset, DatasetFile, RemoteEntity, get_dataset_data_dir, is_dataset_name_valid
-from renku.core.models.provenance.agent import Person
-from renku.core.models.provenance.annotation import Annotation
-from renku.core.utils import communication
-from renku.core.utils.git import clone_repository, get_cache_directory_for_repository, get_git_user
-from renku.core.utils.metadata import is_external_file
-from renku.core.utils.os import get_absolute_path, get_files, get_relative_path, is_subpath
-from renku.core.utils.urls import get_slug, remove_credentials
-
-
-@inject.autoparams("client_dispatcher")
-def create_dataset(
-    name: str,
-    client_dispatcher: IClientDispatcher,
-    title: Optional[str] = None,
-    description: Optional[str] = None,
-    creators: Optional[List[Person]] = None,
-    keywords: Optional[List[str]] = None,
-    images: Optional[List[ImageRequestModel]] = None,
-    update_provenance: bool = True,
-    custom_metadata: Optional[Dict[str, Any]] = None,
-):
-    """Create a dataset."""
-    client = client_dispatcher.current_client
-
-    if not is_dataset_name_valid(name):
-        valid_name = get_slug(name, lowercase=False)
-        raise errors.ParameterError(f'Dataset name "{name}" is not valid (Hint: "{valid_name}" is valid).')
-
-    datasets_provenance = DatasetsProvenance()
-
-    if datasets_provenance.get_by_name(name=name):
-        raise errors.DatasetExistsError(f"Dataset exists: '{name}'")
-
-    if not title:
-        title = name
-
-    if creators is None:
-        creators = [get_git_user(client.repository)]
-
-    keywords = keywords or []
-
-    annotations = None
-
-    if custom_metadata:
-        annotations = [Annotation(id=Annotation.generate_id(), source="renku", body=custom_metadata)]
-
-    dataset = Dataset(
-        identifier=None,
-        name=name,
-        title=title,
-        description=description,
-        creators=creators,
-        keywords=keywords,
-        project_id=client.project.id,
-        annotations=annotations,
-    )
-
-    if images:
-        set_dataset_images(client, dataset, images)
-
-    if update_provenance:
-        datasets_provenance.add_or_update(dataset)
-
-    return dataset
-
-
-def set_dataset_images(client, dataset: Dataset, images: List[ImageRequestModel]):
-    """Set a dataset's images."""
-    if not images:
-        images = []
-
-    image_folder = renku_dataset_images_path(client) / dataset.initial_identifier
-    image_folder.mkdir(exist_ok=True, parents=True)
-    previous_images = dataset.images or []
-
-    dataset.images = []
-    images_updated = False
-    for img in images:
-        img_object = img.to_image_object(dataset)
-
-        if not img_object:
-            continue
-        dataset.images.append(img_object)
-        images_updated = True
-
-    new_urls = [i.content_url for i in dataset.images]
-
-    for prev in previous_images:
-        # NOTE: Delete images if they were removed
-        if prev.content_url in new_urls or urllib.parse.urlparse(prev.content_url).netloc:
-            continue
-
-        path = prev.content_url
-        if not os.path.isabs(path):
-            path = os.path.normpath(os.path.join(client.path, path))
-
-        os.remove(path)
-
-    return images_updated or dataset.images != previous_images
-
-
-def update_dataset_custom_metadata(dataset: Dataset, custom_metadata: Dict):
-    """Update custom metadata on a dataset."""
-
-    existing_metadata = [a for a in dataset.annotations if a.source != "renku"]
-
-    existing_metadata.append(Annotation(id=Annotation.generate_id(), body=custom_metadata, source="renku"))
-
-    dataset.annotations = existing_metadata
-
-
-@inject.autoparams("client_dispatcher")
-def add_data_to_dataset(
-    dataset,
-    urls,
-    client_dispatcher: IClientDispatcher,
-    force=False,
-    overwrite=False,
-    sources: Optional[List[str]] = None,
-    destination="",
-    ref=None,
-    external=False,
-    extract=False,
-    all_at_once=False,
-    destination_names=None,
-    repository: Repository = None,
-    clear_files_before=False,
-):
-    """Import the data into the data directory."""
-    client = client_dispatcher.current_client
-    sources = sources or []
-
-    dataset_datadir = client.path / get_dataset_data_dir(client, dataset)
-    # NOTE: Make sure that dataset's data dir exists because we check for existence of a destination later to decide
-    # what will be its name
-    dataset_datadir.mkdir(parents=True, exist_ok=True)
-
-    destination = destination or Path(".")
-    destination = _resolve_path(dataset_datadir, destination)
-    destination = dataset_datadir / destination
-
-    client.check_external_storage()
-
-    files = []
-    if all_at_once:  # Importing a non-git dataset
-        files = _add_from_urls(urls=urls, destination_names=destination_names, destination=destination, extract=extract)
-    else:
-        for url in urls:
-            is_remote, is_git, url = _check_url(url)
-            if is_git and is_remote:  # Remote repository
-                new_files = _add_from_git(
-                    url=url, sources=sources, destination=destination, ref=ref, repository=repository
-                )
-            else:
-                if sources:
-                    raise errors.UsageError("Cannot use '-s/--src/--source' with URLs or local files.")
-
-                if not is_remote:  # Local path, might be a repository
-                    if is_git:
-                        communication.warn(
-                            "Adding data from local Git repository: Use remote's Git URL instead to enable lineage "
-                            "information and updates."
-                        )
-                    u = urllib.parse.urlparse(url)
-                    new_files = _add_from_local(
-                        client, dataset=dataset, path=u.path, external=external, destination=destination
-                    )
-                else:  # Remote URL
-                    new_files = _add_from_url(client=client, url=url, destination=destination, extract=extract)
-
-            files.extend(new_files)
-
-    # Remove all files that are under a .git directory
-    paths_to_avoid = [f["path"] for f in files if ".git" in str(f["path"]).split(os.path.sep)]
-    if paths_to_avoid:
-        files = [f for f in files if f["path"] not in paths_to_avoid]
-        communication.warn(
-            "Ignored adding paths under a .git directory:\n  " + "\n  ".join(str(p) for p in paths_to_avoid)
-        )
-
-    files_to_commit = {str(client.path / f["path"]) for f in files}
-
-    if not force:
-        ignored_files = client.find_ignored_paths(*files_to_commit)
-        if ignored_files:
-            ignored_files = set(ignored_files)
-            files_to_commit = files_to_commit.difference(ignored_files)
-            ignored_sources = []
-            for file_ in files:
-                if str(client.path / file_["path"]) in ignored_files:
-                    operation = file_.get("operation")
-                    if operation:
-                        src, _, _ = operation
-                        ignored_sources.append(src)
-                    else:
-                        ignored_sources.append(file_["path"])
-
-            files = [f for f in files if str(client.path / f["path"]) in files_to_commit]
-            communication.warn(
-                "Theses paths are ignored by one of your .gitignore "
-                + 'files (use "--force" flag if you really want to add '
-                + "them):\n  "
-                + "\n  ".join([str(p) for p in ignored_sources])
-            )
-
-    # all files at this point can be force-added
-
-    if not overwrite:
-        existing_files = []
-        for path in files_to_commit:
-            relative_path = Path(path).relative_to(client.path)
-            if dataset.find_file(relative_path):
-                existing_files.append(path)
-
-        if existing_files:
-            files_to_commit = files_to_commit.difference(existing_files)
-            files = [f for f in files if str(client.path / f["path"]) in files_to_commit]
-            communication.warn(
-                "These existing files were not overwritten "
-                + '(use "--overwrite" flag to overwrite them):\n  '
-                + "\n  ".join([str(p) for p in existing_files])
-            )
-
-    for data in files:
-        operation = data.pop("operation", None)
-        if not operation:
-            continue
-
-        src, dst, action = operation
-
-        # Remove existing file if any; required as a safety-net to avoid corrupting external files
-        remove_file(dst)
-        dst.parent.mkdir(parents=True, exist_ok=True)
-
-        if action == "copy":
-            shutil.copy(src, dst)
-        elif action == "move":
-            shutil.move(src, dst, copy_function=shutil.copy)
-        elif action == "symlink":
-            _create_external_file(client, src, dst)
-            data["is_external"] = True
-        else:
-            raise errors.OperationError(f"Invalid action {action}")
-
-    # Track non-symlinks in LFS
-    if client.check_external_storage():
-        client.track_paths_in_storage(*files_to_commit)
-
-    # Force-add to include possible ignored files
-    client.repository.add(*files_to_commit, renku_pointers_path(client), force=True)
-
-    n_staged_changes = len(client.repository.staged_changes)
-    if n_staged_changes == 0:
-        communication.warn("No new file was added to project")
-
-    if not files:
-        return
-
-    # Generate the DatasetFiles
-    dataset_files = []
-    for data in files:
-        dataset_file = DatasetFile.from_path(
-            client=client, path=data["path"], source=data["source"], based_on=data.get("based_on")
-        )
-        dataset_files.append(dataset_file)
-
-    if clear_files_before:
-        dataset.clear_files()
-    dataset.add_or_update_files(dataset_files)
-    datasets_provenance = DatasetsProvenance()
-    datasets_provenance.add_or_update(dataset, creator=get_git_user(client.repository))
-
-
-def _add_from_git(client, url, sources, destination, ref, repository: Repository = None):
-    """Process adding resources from another git repository."""
-    from renku.core.management.client import LocalClient
-
-    destination_exists = destination.exists()
-    destination_is_dir = destination.is_dir()
-
-    def check_sources_are_within_remote_repo():
-        for source in sources:
-            if get_relative_path(path=source, base=repository.path) is None:
-                raise errors.ParameterError(f"Path '{source}' is not within the repository")
-
-    def get_paths_from_remote_repo() -> Set[Path]:
-        """Return all paths from the repo that match a source pattern."""
-        if not sources:
-            return set(repository.path.glob("*"))
-
-        paths = set()
-        for source in sources:
-            # NOTE: Normalized source to resolve .. references (if any). This preserves wildcards.
-            normalized_source = os.path.normpath(source)
-            absolute_source = os.path.join(repository.path, normalized_source)
-            # NOTE: Path.glob("root/**") does not return correct results (e.g. it include ``root`` in the result)
-            subpaths = set(Path(p) for p in glob.glob(absolute_source))
-            if len(subpaths) == 0:
-                raise errors.ParameterError("No such file or directory", param_hint=str(source))
-            paths |= subpaths
-
-        return paths
-
-    def get_destination_root(n_paths, path: Path):
-        has_multiple_paths = n_paths > 1
-        multiple_sources = has_multiple_paths or path.is_dir()
-
-        if multiple_sources and destination_exists and not destination_is_dir:
-            raise errors.ParameterError(f"Destination is not a directory: '{destination}'")
-
-        return (
-            destination / path.name
-            if has_multiple_paths or (destination_exists and destination_is_dir)
-            else destination
-        )
-
-    def get_metadata(src, dst) -> Optional[Dict]:
-        path_in_src_repo = src.relative_to(repository.path)
-        path_in_dst_repo = dst.relative_to(client.path)
-
-        if path_in_dst_repo in new_files:  # A path with the same destination is already copied
-            return
-
-        new_files.add(path_in_dst_repo)
-
-        if is_external_file(path=src, client_path=repository.path):
-            operation = (src.resolve(), dst, "symlink")
-        else:
-            operation = (src, dst, "move")
-
-        checksum = repository.get_object_hash(revision="HEAD", path=path_in_src_repo)
-        based_on = RemoteEntity(checksum=checksum, path=path_in_src_repo, url=url)
-
-        return {
-            "path": path_in_dst_repo,
-            "source": remove_credentials(url),
-            "parent": client,
-            "based_on": based_on,
-            "operation": operation,
-        }
-
-    if not repository:
-        repository = clone_repository(
-            url=url,
-            path=get_cache_directory_for_repository(client=client, url=url),
-            checkout_revision=ref,
-            depth=None,
-            clean=True,
-        )
-
-    check_sources_are_within_remote_repo()
-    paths = get_paths_from_remote_repo()
-    n_paths = len(paths)
-
-    LocalClient(repository.path).pull_paths_from_storage(*paths)
-
-    results = []
-    new_files = set()
-    for path in paths:
-        dst_root = get_destination_root(n_paths=n_paths, path=path)
-
-        for file in get_files(path):
-            src = file
-            relative_path = file.relative_to(path)
-            dst = dst_root / relative_path
-
-            metadata = get_metadata(src, dst)
-            if metadata:
-                results.append(metadata)
-
-    return results
-
-
-def _add_from_urls(client, urls, destination, destination_names, extract):
-    """Add files from urls."""
-    if destination.exists() and not destination.is_dir():
-        raise errors.ParameterError(f"Destination is not a directory: '{destination}'")
-
-    destination.mkdir(parents=True, exist_ok=True)
-
-    listeners = communication.get_listeners()
-
-    def subscribe_communication_listeners(function, **kwargs):
-        try:
-            for communicator in listeners:
-                communication.subscribe(communicator)
-            return function(**kwargs)
-        finally:
-            for communicator in listeners:
-                communication.unsubscribe(communicator)
-
-    files = []
-    max_workers = min(os.cpu_count() - 1, 4) or 1
-    with concurrent.futures.ThreadPoolExecutor(max_workers) as executor:
-        futures = {
-            executor.submit(
-                subscribe_communication_listeners,
-                _add_from_url,
-                client=client,
-                url=url,
-                destination=destination,
-                extract=extract,
-                filename=name,
-                multiple=True,
-            )
-            for url, name in zip(urls, destination_names)
-        }
-
-        for future in concurrent.futures.as_completed(futures):
-            files.extend(future.result())
-
-    return files
-
-
-def _add_from_url(client, url, destination, extract, filename=None, multiple: bool = False):
-    """Process adding from url and return the location on disk."""
-    from renku.core.utils import requests
-
-    url = _provider_check(url)
-
-    try:
-        start = time.time() * 1e3
-
-        tmp_root, paths = requests.download_file(
-            base_directory=client.renku_path / CACHE, url=url, filename=filename, extract=extract
-        )
-
-        exec_time = (time.time() * 1e3 - start) // 1e3
-        # If execution time was less or equal to zero seconds,
-        # block the thread a bit to avoid being rate limited.
-        if exec_time == 0:
-            time.sleep(min(os.cpu_count() - 1, 4) or 1)
-    except errors.RequestError as e:  # pragma nocover
-        raise errors.OperationError("Cannot download from {}".format(url)) from e
-
-    paths = [p for p in paths if not p.is_dir()]
-
-    if len(paths) > 1 or multiple:
-        if destination.exists() and not destination.is_dir():
-            raise errors.ParameterError(f"Destination is not a directory: '{destination}'")
-        destination.mkdir(parents=True, exist_ok=True)
-    elif len(paths) == 1:
-        tmp_root = paths[0].parent if destination.exists() else paths[0]
-
-    paths = [(src, destination / src.relative_to(tmp_root)) for src in paths if not src.is_dir()]
-    return [
-        {
-            "operation": (src, dst, "move"),
-            "path": dst.relative_to(client.path),
-            "source": remove_credentials(url),
-            "parent": client,
-        }
-        for src, dst in paths
-    ]
-
-
-def _add_from_local(client, dataset, path, external, destination) -> List[Dict]:
-    """Add a file or directory from a local filesystem."""
-    action = "symlink" if external else "copy"
-    absolute_dataset_data_dir = (client.path / get_dataset_data_dir(client, dataset)).resolve()
-    source_root = Path(get_absolute_path(path))
-    is_with_repo = is_subpath(path=source_root, base=client.path)
-
-    def check_recursive_addition(src: Path):
-        if src.resolve() == absolute_dataset_data_dir:
-            raise errors.ParameterError(f"Cannot recursively add path containing dataset's data directory: {path}")
-
-    def get_destination_root():
-        destination_exists = destination.exists()
-        destination_is_dir = destination.is_dir()
-
-        if client.is_protected_path(source_root):
-            raise errors.ProtectedFiles([source_root])
-
-        check_recursive_addition(source_root)
-
-        if not source_root.exists():
-            raise errors.ParameterError(f"Cannot find source file: {path}")
-        if source_root.is_dir() and destination_exists and not destination_is_dir:
-            raise errors.ParameterError(f"Cannot copy directory '{path}' to non-directory '{destination}'")
-
-        return destination / source_root.name if destination_exists and destination_is_dir else destination
-
-    def get_metadata(src) -> Dict:
-        if is_with_repo:
-            path_in_repo = src.relative_to(client.path)
-            return {"path": path_in_repo, "source": path_in_repo, "parent": client}
-        else:
-            relative_path = src.relative_to(source_root)
-            dst = destination_root / relative_path
-
-            return {
-                "path": dst.relative_to(client.path),
-                "source": os.path.relpath(src, client.path),
-                "parent": client,
-                "operation": (src, dst, action),
-            }
-
-    destination_root = get_destination_root()
-
-    if source_root.is_dir():
-        metadata = []
-        for file in source_root.rglob("*"):
-            if client.is_protected_path(file):
-                raise errors.ProtectedFiles([file])
-
-            if file.is_dir():
-                check_recursive_addition(file)
-                continue
-            metadata.append(get_metadata(file))
-
-        return metadata
-    else:
-        return [get_metadata(source_root)]
-
-
-@inject.autoparams("client_dispatcher", "dataset_gateway")
-def move_files(client_dispatcher: IClientDispatcher, dataset_gateway: IDatasetGateway, files, to_dataset):
-    """Move files and their metadata from one or more datasets to a target dataset."""
-    client = client_dispatcher.current_client
-
-    datasets = [d.copy() for d in dataset_gateway.get_all_datasets()]
-    if to_dataset:
-        # NOTE: Use the same dataset object or otherwise a race happens if dataset is in both source and destination
-        to_dataset: Dataset = next(d for d in datasets if d.name == to_dataset)
-    modified_datasets = {}
-
-    progress_name = "Updating dataset metadata"
-    communication.start_progress(progress_name, total=len(files))
-    try:
-        for src, dst in files.items():
-            src = src.relative_to(client.path)
-            dst = dst.relative_to(client.path)
-            # NOTE: Files are moved at this point, so, we use can use dst
-            new_dataset_file = DatasetFile.from_path(client, dst)
-            for dataset in datasets:
-                removed = dataset.unlink_file(src, missing_ok=True)
-                if removed:
-                    modified_datasets[dataset.name] = dataset
-                    new_dataset_file.based_on = removed.based_on
-                    new_dataset_file.source = removed.source
-                    if not to_dataset:
-                        dataset.add_or_update_files(new_dataset_file)
-
-                # NOTE: Update dataset if it contains a destination that is being overwritten
-                modified = dataset.find_file(dst)
-                if modified:
-                    modified_datasets[dataset.name] = dataset
-                    dataset.add_or_update_files(new_dataset_file)
-
-            if to_dataset:
-                to_dataset.add_or_update_files(new_dataset_file)
-
-            communication.update_progress(progress_name, amount=1)
-    finally:
-        communication.finalize_progress(progress_name)
-
-    datasets_provenance = DatasetsProvenance()
-    for dataset in modified_datasets.values():
-        datasets_provenance.add_or_update(dataset, creator=get_git_user(client.repository))
-    if to_dataset:
-        datasets_provenance.add_or_update(to_dataset, creator=get_git_user(client.repository))
-
-
-@inject.autoparams("client_dispatcher")
-def update_dataset_local_files(client_dispatcher: IClientDispatcher, records: List[DynamicProxy], delete=False):
-    """Update files metadata from the git history."""
-    client = client_dispatcher.current_client
-
-    updated_files: List[DynamicProxy] = []
-    deleted_files: List[DynamicProxy] = []
-    progress_text = "Checking for local updates"
-
-    try:
-        communication.start_progress(progress_text, len(records))
-        check_paths = []
-        records_to_check = []
-
-        for file in records:
-            communication.update_progress(progress_text, 1)
-
-            if file.based_on or file.is_external:
-                continue
-
-            if not (client.path / file.entity.path).exists():
-                deleted_files.append(file)
-                continue
-
-            check_paths.append(file.entity.path)
-            records_to_check.append(file)
-
-        checksums = client.repository.get_object_hashes(check_paths)
-
-        for file in records_to_check:
-            current_checksum = checksums.get(file.entity.path)
-            if not current_checksum:
-                deleted_files.append(file)
-            elif current_checksum != file.entity.checksum:
-                updated_files.append(file)
-    finally:
-        communication.finalize_progress(progress_text)
-
-    if updated_files or (deleted_files and delete):
-        _update_datasets_metadata(client, updated_files, deleted_files, delete)
-
-    return updated_files, deleted_files
-
-
-def _update_datasets_metadata(
-    client,
-    updated_files: List[DynamicProxy],
-    deleted_files: List[DynamicProxy],
-    delete,
-):
-    modified_datasets = {}
-
-    for file in updated_files:
-        new_file = DatasetFile.from_path(
-            client=client, path=file.entity.path, based_on=file.based_on, source=file.source
-        )
-        modified_datasets[file.dataset.name] = file.dataset
-        file.dataset.add_or_update_files([new_file])
-
-    if delete:
-        for file in deleted_files:
-            modified_datasets[file.dataset.name] = file.dataset
-            file.dataset.unlink_file(file.entity.path)
-
-    datasets_provenance = DatasetsProvenance()
-    for dataset in modified_datasets.values():
-        datasets_provenance.add_or_update(dataset, creator=get_git_user(client.repository))
-
-
-@inject.autoparams("client_dispatcher")
-def update_dataset_git_files(client_dispatcher: IClientDispatcher, files: List[DynamicProxy], ref, delete=False):
-    """Update files and dataset metadata according to their remotes.
-
-    :param files: List of files to be updated
-    :param delete: Indicates whether to delete files or not
-
-    :return: List of files that should be deleted
-    """
-    from renku.core.management.client import LocalClient
-
-    client = client_dispatcher.current_client
-
-    visited_repos = {}
-    updated_files: List[DynamicProxy] = []
-    deleted_files: List[DynamicProxy] = []
-
-    progress_text = "Checking files for updates"
-
-    try:
-        communication.start_progress(progress_text, len(files))
-        for file in files:
-            communication.update_progress(progress_text, 1)
-            if not file.based_on:
-                continue
-
-            based_on = file.based_on
-            url = based_on.url
-            if url in visited_repos:
-                remote_repository, remote_client = visited_repos[url]
-            else:
-                remote_repository = clone_repository(
-                    url=url, path=get_cache_directory_for_repository(client=client, url=url), checkout_revision=ref
-                )
-                remote_client = LocalClient(remote_repository.path)
-                visited_repos[url] = remote_repository, remote_client
-
-            checksum = remote_repository.get_object_hash(path=based_on.path, revision="HEAD")
-            found = bool(checksum)
-            changed = found and based_on.checksum != checksum
-
-            src = remote_repository.path / based_on.path
-            dst = client.renku_path.parent / file.entity.path
-
-            if changed:
-                if src.exists():
-                    # Fetch file if it is tracked by Git LFS
-                    remote_client.pull_paths_from_storage(remote_client.path / based_on.path)
-                    if is_external_file(path=src, client_path=remote_client.path):
-                        remove_file(dst)
-                        _create_external_file(client, src.resolve(), dst)
-                    else:
-                        shutil.copy(src, dst)
-                    file.based_on = RemoteEntity(checksum=checksum, path=based_on.path, url=based_on.url)
-                    updated_files.append(file)
-                else:
-                    # File was removed or renamed
-                    found = False
-
-            if not found:
-                if delete:
-                    remove_file(dst)
-                deleted_files.append(file)
-    finally:
-        communication.finalize_progress(progress_text)
-
-    if not updated_files and (not delete or not deleted_files):
-        # Nothing to commit or update
-        return [], deleted_files
-
-    # Commit changes in files
-
-    file_paths = {str(client.path / f.entity.path) for f in updated_files + deleted_files}
-    # Force-add to include possible ignored files that are in datasets
-    client.repository.add(*file_paths, force=True)
-
-    _update_datasets_metadata(client, updated_files, deleted_files, delete)
-
-    return updated_files, deleted_files
-
-
-def update_external_files(client, records: List[DynamicProxy]):
-    """Update files linked to external storage."""
-    updated_files_paths = []
-    updated_datasets = {}
-
-    for file in records:
-        if file.is_external:
-            path = client.path / file.entity.path
-            link = path.parent / os.readlink(path)
-            pointer_file = client.path / link
-            pointer_file = _update_pointer_file(client, pointer_file)
-            if pointer_file is not None:
-                relative = os.path.relpath(pointer_file, path.parent)
-                os.remove(path)
-                os.symlink(relative, path)
-                updated_files_paths.append(str(path))
-                updated_datasets[file.dataset.name] = file.dataset
-
-    if not updated_files_paths:
-        return
-
-    client.repository.add(*updated_files_paths, force=True)
-    client.repository.add(renku_pointers_path(client), force=True)
-
-    datasets_provenance = DatasetsProvenance()
-
-    for dataset in updated_datasets.values():
-        for file in dataset.files:
-            if str(client.path / file.entity.path) in updated_files_paths:
-                new_file = DatasetFile.from_path(client=client, path=file.entity.path, source=file.source)
-                dataset.add_or_update_files(new_file)
-
-        datasets_provenance.add_or_update(dataset, creator=get_git_user(client.repository))
-
-
-def _create_pointer_file(client, target, checksum=None):
-    """Create a new pointer file."""
-    target = Path(target).resolve()
-
-    if checksum is None:
-        checksum = _calculate_checksum(client.repository, target)
-        assert checksum is not None, f"Cannot calculate checksum for '{target}'"
-
-    while True:
-        filename = f"{uuid.uuid4()}-{checksum}"
-        path = renku_pointers_path(client) / filename
-        if not path.exists():
-            break
-
-    try:
-        os.symlink(target, path)
-    except FileNotFoundError:
-        raise errors.ParameterError("Cannot find external file {}".format(target))
-
-    return path
-
-
-def _update_pointer_file(client, pointer_file_path):
-    """Update a pointer file."""
-    try:
-        target = pointer_file_path.resolve(strict=True)
-    except FileNotFoundError:
-        target = pointer_file_path.resolve()
-        raise errors.ParameterError("External file not found: {}".format(target))
-
-    checksum = _calculate_checksum(client.repository, target)
-    current_checksum = pointer_file_path.name.split("-")[-1]
-
-    if checksum == current_checksum:
-        return
-
-    os.remove(pointer_file_path)
-    return _create_pointer_file(client, target, checksum=checksum)
-
-
-def remove_file(filepath):
-    """Remove a file/symlink and its pointer file (for external files)."""
-    path = Path(filepath)
-    try:
-        link = path.parent / os.readlink(path)
-    except FileNotFoundError:
-        return
-    except OSError:  # not a symlink but a normal file
-        os.remove(path)
-        return
-
-    os.remove(path)
-
-    try:
-        os.remove(link)
-    except FileNotFoundError:
-        pass
-
-
-def _calculate_checksum(repository, filepath):
-    try:
-        return repository.hash_objects([filepath])[0]
-    except errors.GitCommandError:
-        raise
-
-
-def _create_external_file(client, src, dst):
-    """Create a new external file."""
-    try:
-        pointer_file = _create_pointer_file(client, target=src)
-        relative = os.path.relpath(pointer_file, dst.parent)
-        os.symlink(relative, dst)
-    except OSError as e:
-        raise errors.OperationError("Could not create symbolic link") from e
-
-
-def _ensure_dropbox(url):
-    """Ensure dropbox url is set for file download."""
-    if not isinstance(url, urllib.parse.ParseResult):
-        url = urllib.parse.urlparse(url)
-
-    query = url.query or ""
-    if "dl=0" in url.query:
-        query = query.replace("dl=0", "dl=1")
-    else:
-        query += "dl=1"
-
-    url = url._replace(query=query)
-    return url
-
-
-def _provider_check(url):
-    """Check additional provider related operations."""
-    from renku.core.utils import requests
-
-    url = requests.get_redirect_url(url)
-    url = urllib.parse.urlparse(url)
-
-    if "dropbox.com" in url.netloc:
-        url = _ensure_dropbox(url)
-
-    return urllib.parse.urlunparse(url)
-
-
-def _resolve_path(root_path, path):
-    """Check if a path is within a root path and resolve it."""
-    try:
-        # NOTE: Path.resolve also works with paths with wildcards
-        root_path = Path(root_path).resolve()
-        path = os.path.abspath(root_path / path)
-        return Path(path).relative_to(root_path)
-    except ValueError:
-        raise errors.ParameterError("File {} is not within path {}".format(path, root_path))
-
-
-def _check_url(url):
-    """Check if a url is local/remote and if it contains a git repository."""
-    from renku.core.utils import requests
-
-    u = urllib.parse.urlparse(url)
-
-    if u.scheme not in SUPPORTED_SCHEMES:
-        raise errors.UrlSchemeNotSupported('Scheme "{}" not supported'.format(u.scheme))
-
-    is_remote = u.scheme not in ("", "file") or url.startswith("git@")
-    is_git = False
-
-    if is_remote:
-        is_git = u.path.endswith(".git")
-        if not is_git:
-            url = requests.get_redirect_url(url)
-    elif os.path.isdir(u.path) or os.path.isdir(os.path.realpath(u.path)):
-        try:
-            Repository(u.path, search_parent_directories=True)
-        except errors.GitError:
-            pass
-        else:
-            is_git = True
-
-    return is_remote, is_git, url
