+++ renku/core/utils/os.py
@@ -34,18 +34,21 @@ def get_relative_path_to_cwd(path: Union[Path, str]) -> str:
 def get_absolute_path(path: Union[Path, str], base: Union[Path, str] = None) -> str:
     """Return absolute normalized path without resolving symlinks."""
     if base is not None:
+        base = Path(base).resolve()
         path = os.path.join(base, path)
 
     # NOTE: Do not use os.path.realpath or Path.resolve() because they resolve symlinks
     return os.path.abspath(path)
 
 
-def get_relative_path(path: Union[Path, str], base: Union[Path, str]) -> Optional[Path]:
+def get_relative_path(path: Union[Path, str], base: Union[Path, str], strict: bool = False) -> Optional[Path]:
     """Return a relative path to the base if path is within base without resolving symlinks."""
     try:
         absolute_path = get_absolute_path(path=path, base=base)
         return Path(absolute_path).relative_to(base)
     except ValueError:
+        if strict:
+            raise errors.ParameterError("File {} is not within path {}".format(path, base))
         return
 
 
@@ -117,10 +120,27 @@ def normalize_to_ascii(input_string, sep="-"):
     )
 
 
-def delete_file(path: Union[Path, str], ignore_errors: bool = True):
-    """Delete a file."""
+def delete_file(filepath: Union[Path, str], ignore_errors: bool = True, follow_symlinks: bool = False):
+    """Remove a file/symlink and its pointer file (for external files)."""
+    path = Path(filepath)
+    link = None
     try:
-        os.unlink(path)
+        link = path.parent / os.readlink(path)
+    except FileNotFoundError:
+        if not ignore_errors:
+            raise
+        return
+    except OSError:  # not a symlink but a normal file
+        pass
+
+    try:
+        os.remove(path)
     except OSError:
         if not ignore_errors:
             raise
+
+    if follow_symlinks and link:
+        try:
+            os.remove(link)
+        except FileNotFoundError:
+            pass
