+++ tests/core/commands/test_dataset.py
@@ -26,11 +26,17 @@ from pathlib import Path
 import pytest
 
 from renku.core import errors
-from renku.core.commands.dataset import add_to_dataset, create_dataset_command, file_unlink, list_datasets, list_files
+from renku.core.commands.dataset import (
+    add_to_dataset_command,
+    create_dataset_command,
+    file_unlink_command,
+    list_datasets_command,
+    list_files_command,
+)
 from renku.core.errors import ParameterError
 from renku.core.management.dataset.context import DatasetContext
+from renku.core.management.dataset.dataset_add import add_data_to_dataset
 from renku.core.management.dataset.datasets_provenance import DatasetsProvenance
-from renku.core.management.dataset.usecase import add_data_to_dataset
 from renku.core.management.repository import DEFAULT_DATA_DIR as DATA_DIR
 from renku.core.metadata.repository import Repository
 from renku.core.models.dataset import Dataset, is_dataset_name_valid
@@ -59,16 +65,14 @@ def test_data_add(scheme, path, overwrite, error, client_with_injection, directo
         if path == "temp":
             path = str(directory_tree / "file1")
 
-        with DatasetContext(name="dataset", create=True, commit_database=True) as d:
-            d.creators = [Person(name="me", email="me@example.com", id="me_id")]
-            add_data_to_dataset(d, ["{}{}".format(scheme, path)], overwrite=overwrite)
+        dataset = add_data_to_dataset("dataset", ["{}{}".format(scheme, path)], overwrite=overwrite, create=True)
 
         target_path = os.path.join(DATA_DIR, "dataset", "file1")
 
         with open(target_path) as f:
             assert f.read() == "file1 content"
 
-        assert d.find_file(target_path)
+        assert dataset.find_file(target_path)
 
         # check that the imported file is read-only
         assert not os.access(target_path, stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH)
@@ -76,19 +80,16 @@ def test_data_add(scheme, path, overwrite, error, client_with_injection, directo
         # check the linking
         if scheme in ("", "file://"):
             shutil.rmtree("./data/dataset")
-            with DatasetContext(name="dataset") as d:
-                d.creators = [Person(name="me", email="me@example.com", id="me_id")]
-                add_data_to_dataset(d, ["{}{}".format(scheme, path)], overwrite=True)
+            dataset = add_data_to_dataset("dataset", ["{}{}".format(scheme, path)], overwrite=True)
+
             assert os.path.exists(target_path)
 
 
 def test_data_add_recursive(directory_tree, client_with_injection):
     """Test recursive data imports."""
-    with DatasetContext(name="dataset", create=True) as dataset:
-        dataset.creators = [Person(name="me", email="me@example.com", id="me_id")]
-        add_data_to_dataset(dataset, [str(directory_tree / "dir1")])
+    dataset = add_data_to_dataset("dataset", [str(directory_tree / "dir1")], create=True)
 
-        assert os.path.basename(os.path.dirname(dataset.files[0].entity.path)) == "dir1"
+    assert os.path.basename(os.path.dirname(dataset.files[0].entity.path)) == "dir1"
 
 
 def test_creator_parse():
@@ -134,7 +135,7 @@ def test_list_datasets_default(project):
         "ds1", title="", description="", creators=[]
     )
 
-    datasets = list_datasets().with_database().build().execute().output
+    datasets = list_datasets_command().with_database().build().execute().output
 
     assert isinstance(datasets, list)
     assert "ds1" in [dataset.title for dataset in datasets]
@@ -148,8 +149,8 @@ def test_list_files_default(project, tmpdir):
     data_file = tmpdir / Path("some-file")
     data_file.write_text("1,2,3", encoding="utf-8")
 
-    add_to_dataset().build().execute([str(data_file)], "ds1")
-    files = list_files().build().execute(datasets=["ds1"]).output
+    add_to_dataset_command().build().execute("ds1", [str(data_file)])
+    files = list_files_command().build().execute(datasets=["ds1"]).output
 
     assert isinstance(files, list)
     assert "some-file" in [Path(f.entity.path).name for f in files]
@@ -159,10 +160,12 @@ def test_unlink_default(directory_tree, client):
     """Test unlink default behaviour."""
     with chdir(client.path):
         create_dataset_command().with_database(write=True).build().execute("dataset")
-        add_to_dataset().with_database(write=True).build().execute([str(directory_tree / "dir1")], "dataset")
+        add_to_dataset_command().with_database(write=True).build().execute(
+            dataset_name="dataset", urls=[str(directory_tree / "dir1")]
+        )
 
     with pytest.raises(ParameterError):
-        file_unlink().build().execute("dataset", (), ())
+        file_unlink_command().build().execute("dataset", (), ())
 
 
 @pytest.mark.xfail
